requirements.txt
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.3.0
colorama==0.4.3
distlib==0.3.0
filelock==3.0.12
importlib-metadata==1.5.0
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
six==1.14.0
virtualenv==20.0.4
wcwidth==0.1.8
zipp==3.0.0
scipy>=0.15.1
numpy>=1.9.2
pytest-flask==1.0.0